python/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
ax253==0.1.5.post1
bidict==0.23.1
bitarray==2.9.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
keyboard==0.13.5
kiss3==8.0.0
multidict==6.0.5
pyham_ax25==1.0.0
pyserial==3.5
pyserial-asyncio==0.6
python-engineio==4.9.1
setuptools==69.5.1
simple-websocket==1.0.0
wheel==0.43.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
